pyproject.toml
[tool.poetry]
name = "MIND"
version = "1.1.0"
description = "motion insights and diagnostics"
authors = ["MMAGJOGE <Joel.ExternGeiser@maxongroup.com>", "JOELGEISER <Joel.Geiser@noser.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.30"
pandas = "^2.2.2"
pydantic-settings = "^2.10.0"
plotly = "^6.1.2"
catboost = "^1.2.5"
onnxruntime = "^1.22.0"
skl2onnx = "^1.17.0"
paho-mqtt = "^2.1.0"
statsmodels = "^0.14.2"
python-jose = "^3.3.0"
bcrypt = "^4.1.3"
asyncpg = "^0.30.0"
aiosqlite = "^0.20.0"
libsass = "^0.23.0"
seaborn = "^0.13.2"
kneed = "^0.8.5"
tqdm = "^4.67.1"
prometheus-client = "^0.22.0"
psutil = "^7.0.0"
alembic = "^1.14.1"
uuid7 = "^0.1.0"
scipy = "1.16.0"
fastapi = {extras = ["standard"], version = "^0.115.13"}
pywavelets = "^1.8.0"
numpy = "2.3.1"
gunicorn = "^23.0.0"
pydantic = "^2.11.7"
onnx = "^1.18.0"
matplotlib = "^3.10.3"
scikit-learn = "^1.7.0"
pyyaml = "^6.0.2"
httpx = "^0.28.1"
ruamel-yaml = "^0.18.15"
cachetools = "^6.2.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
sqlalchemy-utils = "^0.41.2"
pytest-cov = "^6.0.0"
pytest-mock = "^3.14.0"
pytest-env = "^1.1.3"
watchfiles = "^1.1.0"
pytest-asyncio = "^1.0.0"


[[tool.poetry.source]]
name = "pytorch_cpu"
url = "http://download.pytorch.org/whl/cpu"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"